+++ packages/network/test/integration/network-stabilization.test.ts
@@ -4,7 +4,7 @@ import assert from 'assert'
 
 import { wait } from '@streamr/utils'
 
-import { createNetworkNode } from '../../src/composition'
+import { createNetworkNode } from '../../src/createNetworkNode'
 import { StreamPartIDUtils } from '@streamr/protocol'
 
 function areEqual(a: any, b: any) {
