+++ packages/network/test/integration/webrtc-error-scenarios-during-signalling.test.ts
@@ -2,7 +2,7 @@ import { runAndWaitForEvents } from '@streamr/test-utils'
 
 import { Tracker, startTracker } from '@streamr/network-tracker'
 import { NetworkNode } from '../../src/logic/NetworkNode'
-import { createNetworkNode } from '../../src/composition'
+import { createNetworkNode } from '../../src/createNetworkNode'
 import { Event as NodeEvent } from '../../src/logic/Node'
 import { Event as NodeToTrackerEvent } from '../../src/protocol/NodeToTracker'
 import { toStreamID, toStreamPartID } from '@streamr/protocol'
