+++ packages/protocol/test/unit/protocol/control_layer/ProxyConnectionRequestSerializerV2.test.ts
@@ -1,6 +1,8 @@
 import assert from 'assert'
-
-import { ProxyConnectionRequest, ControlMessage, toStreamID } from '../../../../src/index'
+import ControlMessage from '../../../../src/protocol/control_layer/ControlMessage'
+import ProxyConnectionRequest from '../../../../src/protocol/control_layer/proxy_connection_request/ProxyConnectionRequest'
+import '../../../../src/protocol/control_layer/proxy_connection_request/ProxyConnectionRequestSerializerV2'
+import { toStreamID } from '../../../../src/utils/StreamID'
 import { ProxyDirection } from '../../../../src/utils/types'
 
 const VERSION = 2
