+++ packages/protocol/test/unit/protocol/control_layer/UnsubscribeRequestSerializerV2.test.ts
@@ -1,6 +1,8 @@
 import assert from 'assert'
-
-import { UnsubscribeRequest, ControlMessage, toStreamID } from '../../../../src/index'
+import ControlMessage from '../../../../src/protocol/control_layer/ControlMessage'
+import UnsubscribeRequest from '../../../../src/protocol/control_layer/unsubscribe_request/UnsubscribeRequest'
+import '../../../../src/protocol/control_layer/unsubscribe_request/UnsubscribeRequestSerializerV2'
+import { toStreamID } from '../../../../src/utils/StreamID'
 
 const VERSION = 2
