+++ packages/protocol/test/unit/protocol/tracker_layer/InstructionMessageSerializerV2.test.ts
@@ -1,7 +1,8 @@
 import assert from 'assert'
-
-import { InstructionMessage, toStreamID } from '../../../../src'
+import InstructionMessage from '../../../../src/protocol/tracker_layer/instruction_message/InstructionMessage'
+import '../../../../src/protocol/tracker_layer/instruction_message/InstructionMessageSerializerV2'
 import TrackerMessage from '../../../../src/protocol/tracker_layer/TrackerMessage'
+import { toStreamID } from '../../../../src/utils/StreamID'
 
 const VERSION = 2
